get rid of parent interface and no-op implementation
