--- tinkerpop3/src/main/java/io/shiftleft/overflowdb/structure/NoOpReferenceManager.java
@@ -1,19 +1,0 @@
-package io.shiftleft.overflowdb.structure;
-
-public class NoOpReferenceManager implements ReferenceManager {
-  @Override
-  public void registerRef(NodeRef ref) {
-  }
-
-  @Override
-  public void applyBackpressureMaybe() {
-  }
-
-  @Override
-  public void clearAllReferences() {
-  }
-
-  @Override
-  public void close() {
-  }
-}
